backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
google-genai
python-dotenv
sqlalchemy
alembic
pydantic-settings
youtube-transcript-api==1.2.3
requests
feedparser
APScheduler
pytest
httpx
pytest-mock
gunicorn
beautifulsoup4
lxml
